fix(cvc): FaultMask must be uint16 for 12-bit Com signal
Root cause: Vehicle_State_FaultMask is 12 bits in DBC (BitSize=12).
Com_SendSignal for BitSize>8 reads uint16* from the data pointer.
But faultMask was declared uint8, so the uint16 read picked up stack
garbage in the high byte, corrupting the neighboring Vehicle_State_Mode
signal (4-bit at bit 16, sharing byte 2 with FaultMask low bits).

Result: Mode=0 was overwritten with garbage, appearing as Mode=2 on CAN
even though VSM was in INIT(0). Heartbeat (separate PDU, separate signal)
correctly showed Mode=0 since its OperatingMode was 4-bit = uint8.

Fix: declare faultMask as uint16 to match the 12-bit DBC signal width.

+++ firmware/ecu/cvc/src/Swc_CvcCom.c
@@ -265,7 +265,7 @@ void Swc_CvcCom_TransmitSchedule(uint32 currentTimeMs)
         uint8  txBuf[8];
         uint8  j;
         uint32 faultSig;
-        uint8  faultMask = 0u;
+        uint16 faultMask = 0u;  /* 12-bit signal → must be uint16 for Com_SendSignal */
 
         for (j = 0u; j < 8u; j++) { txBuf[j] = 0u; }
